Worked around DEFINE_SEMAPHORE() API change in latest kernel snapshot
An argument was added to DECLARE_SEMAPHORE() by
https://lore.kernel.org/lkml/20230405203505.1343562-2-mcgrof@kernel.org/
and there's no simple way at the moment to continue using that
macro and still build on both release and snapshot kernels. This
change temporarily makes our only use of it a lower-level macro
that's the same in both kernel versions.

Pair: joe

.../RequestQueue_n1.c
Temporarily switched from DEFINE_SEMAPHORE() to
  __SEMAPHORE_INITIALIZER() (essentially inlining the former).

[git-p4: depot-paths = "//eng/linux-vdo/": change = 158196]