Made recovery more like rebuild [VDOUP-6].
Synthesizing decrefs required playing into the slab journals before
recovering the block map since the latter could destroy the information
needed for the synthesis. Read-only rebuild, however, requires recoving
the block map first as it rebuilds the reference counts from it. Now
that we no longer synthesize decrefs, this change modifies recovery to
also recover the block map first, making rebuild and recovery more
similar.

This change also eliminates the need for the replaying vdo state. This
state was needed to handle the case where the slab journals had been
recovered, but then the vdo crashed before the block map recovery was
complete. With synthesized decrefs, it was necessary to avoid
synthesizing them twice as this could cause reference count
underflow. The replaying state was used to indicate to recovery that
the slab journal replay should be skipped, but that the block map
recovery was still needed. Now that we don't synthesize decrefs, this
state is obsolete. Furthermore, with this change, the block map is now
recovered first, and it is always safe to do this again before
recovering the slab journals. In the event of a vdo which was in the
replaying state before upgrade, it will also have an old format
journal, which will result in the vdo going into read-only mode on
recovery, replacing the replaying state with read-only.

Pair: msakai

.../recovery.c
Added launch_recover_completion() as sugar for
  prepare_recovery_completion() followed immediately by
  vdo_launch_completion().
Moved loading of the slab depot to a new callback, load_slab_depot()
  which is called after the block map replay for all forms of repair.
Modified recover_block_map() to be used by both rebuild and recovery
  immediately after parsing the journal.
Removed is_replaying() and the checking of it since this change
  eliminates the need to put the vdo into the replaying state. Also
  eliminated the super block save and launch of block map recovery
  following slab journal replay.
Renamed extract_increments() to extract_new_mappings() as this function
  is used to find the entries to be applied to the block map.
Added parse_journal() as the single bottle-neck for examining the
  journal. Respelled the separate parsing for recover vs rebuild as
  parse_journal_for_recovery() and parse_journal_for_rebuild(). These
  are called from parse_journal() after it calls
  find_recovery_journal_head_and_tail() directly.
Modified finish_journal_load() to call parse_journal() and then recover
  the block map for both rebuild and recovery.

.../recovery-journal.h
.../recovery-journal.c
.../recovery.c
Renamed vdo_initialize_recovery_journal_post_rebuild() to
  vdo_initialize_recovery_journal_post_repair() and modified
  finish_recovery() to use it for both rebuild and recovery.
Modified vdo_replay_into_slab_journals() to initialize the recovery
  completion's logical_blocks_used and block_map_data_blocks fields
  from the values recorded in the super block rather than setting them
  separately in the journal on the recovery path.

.../types.h
Added comment explaining that VDO_REPLAYING is obsolete.

.../SlabJournalRecovery_t1.c
Removed the check that replaying into the slab journals leaves the
  block map unchanged since we now recover the block map first.

.../Rebuild_t1.c
Replaced testCrashBeforeBlockMapRebuild() which, with the reordering of
  recovery, is the same as crashing before the super block write at the
  end of recovery, with two new tests:
- testCrashBeforeRefCountRebuild() crashes the vdo after recovering the
  block map and then does a full recovery.
- testErrorBeforeRefCountRebuild() injects an error in the slab depot
  load just after recovering the block map, and then tests that the
  ensuing read-only vdo can be rebuilt.
Renamed testCrashDuringRebuild() to testCrashDuringRecovery() since
  that is what it actually does.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 158207]