Reorder recovery.c to reflect reordering of recovery [VDOUP-6].
Change 158207 made rebuild and recovery more similar by re-ordering the
recovery process to recover the block map before recovering the
reference counts as rebuild does. For ease of review, that change did
as little code motion as possible. This change is pure code motion to
reorder the code to reflect the reordering of the process.

Pair: sclafani

.../recovery.c
Reordered to reflect control flow.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 158210]