Move murmurhash3.[hc] into the vdo source directory.
For upstreaming, MurmurHash3 is part of VDO. In the future, if
there is demand, it can be added to the kernel library. This
also allows some simplification of the Makefiles for building
test modules.

Pair: msakai

.../linux/murmurhash3.h
.../murmurhash3.h
Move murmurhash3.h to the uds directory, from which it will be
  copied to the dm-vdo directory.

.../murmur/murmurhash3.c
.../murmurhash3.c
Move murmurhash3.c to the uds directory, from which it will be
  copied to the dm-vdo directory.
Remove module machinery from murmurhash3.c because it is not
  built as a separate module at this time.

.../uds/defines
.../c++/uds/kernelLinux/build/Makefile
.../packaging/kpatch/Makefile
.../MANIFEST.yaml
Eliminate special handling of murmurhash3.[hc].

.../kernelLinux/defines
.../Makefile.module.in
Eliminate the no longer needed UDS_MURMURHASH3 preprocessor macro.

.../userLinux/build/Makefile
Remove the unused vpath directives for the murmur and util
  directories.

.../packaging/src-dist/user/utils/uds/Makefile
Remove the unused vpath directive for the murmur directory.
Alphabetize the list of objects.

.../uds-module.c
Export the murmurhash3_128 symbol for use by the kernel unit
  tests.

.../uds/src/tests/MurmurHash3_t1.c
.../vdo/tests/MurmurHash3_t1.c
.../fillBufferFromSeed.c
.../testPrototypes.h
.../index-layout.c
.../data-vio.c
.../dedupe.c
.../MurmurHash3_p1.c
.../vdoFillIndex.c
Change to include murmurhash3.h locally.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 158226]