Configure fio without local-cpu-specific optimizations.
By default fio will try to build on x86 using "-march=native"
which instructs the compiler to enable instruction set
extensions available on the local processor. Since we have a
mix of Intel CPU generations, this can produce an executable
that won't run on other of our machines.

In particular, this can result in BMI2 shift instructions
being used because they were available on a build machine, and
then running on a different machine without BMI2 will crash
with an illegal instruction fault.

The --disable-optimizations flag to configure tells it not to
use this compiler option. The Fedora spec file uses this
option.

Pair: jwiele

.../Makefile
When configuring, pass --disable-optimizations.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 158359]