Proof of concept: Break up patch series into smaller patches.
This generates a series of patches by using separate YAML
manifest files to add the files incrementally. The
patchset/Makefile iterates over a list of "sub-patches",
running prepareSourceDist.pl on each manifest, then copying
the prepared files to the kernel tree and committing
them. Once all the files are commited, the branch description
is updated from the cover letter and git format-patch is
called to generate the patch series.

In addition, dedupe.c is added in two parts by using an #ifdef
to partition the source. The first prepare pass undefines the
PART2 macro and the second pass defines it, so the first pass
adds a skinny version of dedupe.c and the second pass adds the
whole dedupe.c file. Thus the second diff is the addition of
the sources wrapped with #ifdef PART2. It might be better to
invert the sense of the #ifdef.

Pair: msakai

.../patchset/Makefile
Process multiple patches.

.../dedupe.c
Insert #ifdef PART2 markers for part 2 of the kernel patch for
  this file: the first part of the patch will be the file
  minus the part enclosed in #ifdef PART2; the second part of
  the patch is the remainder.

.../cover.yaml
.../dedupe1.yaml
.../dedupe2.yaml
.../doc.yaml
.../murmur.yaml
.../queue.yaml
.../target.yaml
.../threads.yaml
.../uds.yaml
.../vdo.yaml
Spccify the contents of the smaller patches.

.../murmur.txt
.../queue.txt
.../threads.txt
.../dedupe1.txt
.../dedupe2.txt
The commit messages that describe the smaller patches.

.../vdo/base/Makefile
.../Makefile.module.in
.../src-dist/MANIFEST.yaml
Define PART2 so that the whole file with PART2 markers will be
  built.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 158369]