Generate kpatch against the correct branch.
When generating a kernel patch against a specific branch, such
as the for-next branch in the linux-dm repository, the kernal
patch needs to be generated relative to the specified
branch. Generating a patch against the default branch will
likely include many irrelevant commits.

Pair: msakai

.../kpatch/Makefile
Generate the kpatch against the specified branch, rather than
  the default branch. 

[git-p4: depot-paths = "//eng/linux-vdo/": change = 158396]