Fixed kernel version check workaround in uds-threads.c
Change 155452 for VDO-5368 could not distinguish between kernel
versions with and without an API change at the time, so resorted
to a temporary workaround of checking for symbols associated with
the new function. It's now possible to use our usual boilerplate
for checking kernel versions that gets ifdef'ed out properly for
upstreaming.

Pair: msakai

.../uds-threads.c
Replaced workaround with proper kernel version checks.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 158407]