[ALB-3116] Add a short delay before removing console monitor.
In consoleMonitorTest the time between posting the message, via
pr_err(), and tearing down the console monitor can occur faster than
the print thread handling the message actually does so.  This makes it
appear that the message does not make it to the console when it fact
it does.

This change adds a short delay, which is sufficient to addresses the
issue, to give the thread time to process the message.

Pair: msakai

.../linux-vdo/src/c++/uds/kernelLinux/tests/RequestQueue_n1.c
In consoleMonitorTest added a short delay after posting the message
  before tearing down the console monitor.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 158422]