Adjust srpm packaging manifests to fix sysfs test failures. [VDO-5542]
In our perl tests, multiple Sysfs.pm subtests fail due to the kmod-kvdo
rpm not being compiled with VDO_INTERNAL when rpmbuild compiles it from
the source rpm. This change fixes that and adjusts the packaging
manifests to include/exclude the appropriate files.

Pair: msakai 

.../rpm/kernel/Makefile
Add a new makefile that includes the necessary internal compilation
  flags for when rpmbuild builds the rpms during our perl tests.

.../rpm/MANIFEST.yaml
Adjust the manifest to use the new makefile instead of the src-dist one.
Remove the histogram.[ch] exclusions, as those files are not included
  in the user space source file list.

.../src-dist/MANIFEST.yaml
Add vdo-histograms.[ch] to the manifest, as it is needed when building
  the kmod-kvdo rpm from the srpm during our perl tests.

.../rpm/MANIFEST.yaml
.../github/MANIFEST.yaml
.../kpatch/MANIFEST.yaml
.../patchset/MANIFEST.yaml
Add vdo-histograms.[ch] to the list of exclusions for the tarball(s).

[git-p4: depot-paths = "//eng/linux-vdo/": change = 158457]