Make nfsclientClass the same as clientClass to prevent deadlock.
This prevents a deadlock situation when clientClass and
nfsclientClass are equivalent but not equal. For instance if
they refer to the same distro but one contains PFARM and the
other contains FARM, they will be reserved with separate calls
to rsvpd, but from the same pool. If several tests reserve
machines using clientClass, then try to reserve additional
machines using an equivalent nfsclientClass, they may wait
indefinitely because no test can reserve the additional
machine needed to make progress to free any machines.

Pair: msakai

.../VDOTest.pm
Force nfsclientClass to be the same as clientClass if
  nfsclientClass is not explicitly specified so that both can
  be reserved in a single call the the rsvpd server.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 158500]