Don't define MODULE_VERSION when generating patchsets.
Other dm targets don't define a module version. We use it to record the
CURRENT_VERSION, but a monotonically increasing version is hard to
define when we are upstream since patches are not guaranteed to be
accepted in the order in which they were submitted. Also in the kernel,
this version is redundant with the kernel version.

Pair: msakai

.../dm-vdo-target.c
Wrap MODULE_VERSION in new RHEL_INTERNAL ifdef.

.../Makefile
.../Makefile.module.in
.../src-dist/MANIFEST.yaml
Define RHEL_INTERNAL.

.../patchset/MANIFEST.yaml
Undefine RHEL_INTERNAL.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 158548]