Recapitulate changes 157232 and 157356 on linux-vdo.
Note that the scenario names on linux-vdo specify the OS name
(i.e. '_RHEL9_' in the scenario name). This is to account for
linux-vdo supporting RHEL and Fedora hosts. Because the relevant
scenarios have already been added to the file, the merge of
SupportedScenarios.yaml is a no-op merge. 

Pair: Scliff 

.../MultiArchUpgrade.pm
.../Upgrade.pm
.../MigrationBase.pm
.../SupportedScenarios.yaml
.../vdotests.suites
157232: Add a multi-arch upgrade test and suite to //eng/vdo.
157356: Remove ppc64le from the multi-arch test scenario list.

.../MigrationBase.pm
Modify reserveHosts() to separate the call to the RSVPer 
  getHostOSArchClasses method from the sort operation. This is
  needed because getHostOSArchClasses() can return undef, thus
  causing the sort to fail.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 158594]