Correct ISCSI activation failure on FEDORA37 hosts. [VDO-5563]
Change 158684 ensured that lsscsi is installed on Fedora37+ hosts.
Now that an ISCSI device is found, a subsequent failure was revealed
where lvmdevices is not included in lvm version 2.03.11-9 used on our
lab F37 hosts. As such, the ISCSI activation procedure needs to be
modified for cases where lvmdevices is not found.

Pair: msakai 

.../ISCSI.pm
Modify activate() to skip the add device and delete device logic
  when lvmdevices is not found.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 158697]