Only set FUA on REQ_OP_WRITE bios in fua driver. [VDO-5558]
As of kernel 6.3.11 the kernel checks the operation of
FUA/PREFLUSH bios and rejects operations it's not expecting,
and comments in the commit indicate that WRITE and ZONE_APPEND
are the only operations expected by the kernel flush code. So
DISCARD+FUA is rejected.

Pair: msakai

.../dmFua.c
Only set FUA on REQ_OP_WRITE bios.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 158743]