Make BlockDevice03 work on PMI machines.
BlockDevice03 assumes that vdoFillIndex will write as many index
entries as fit in the index. However, if the zone count is greater
than 1, each chapter will lose a few entries due to zones closing
early, and the test's expectations are not met.

Pair: sclafani

.../vdoFillIndex.c
Set the zone count to always be 1.	

[git-p4: depot-paths = "//eng/linux-vdo/": change = 158766]