Rewrite Suspend tests to work with vdoFillIndex. [VDO-5480]
Change 159039 removed the old-style fill-index dmsetup message in favor
using of the newer vdoFillIndex user tool instead. However,
the specific way the the Suspend tests used fill-index did not work
with vdoFillIndex because the latter only works if the VDO volume is
stopped when it is run. This change adds a new option to vdoFillIndex
and rewrites the Suspend tests to use it to produce the same
preconditions that the tests want, namely a VDO with a short recovery
journal but a long UDS index recovery.

Pair:  sclafani

.../vdoFillIndex.c
Added a new --force-rebuild option to vdoFillIndex that skips closing
  the index. This causes the index to attempt to do a rebuild the next
  time it is started.

.../VDO.pm
Moved the call to dumpConfig() from fillIndex() into its callers
  because it requires the VDO device to be started. This allows the
  VDO device to be stopped when fillIndex() is called, as long as the
  index config was collected earlier.
Updated fillIndex() to use runWhileStopped() so the VDO device
  doesn't have to be started when fillIndex() is called.
Added an optional parameter for the new force-rebuild option.

.../Suspend.pm
Inlined causeDoryRecovery() into makeRecoveringVDO() in order
  to call fillIndex() while the VDO is stopped, before recovery.
  Also simplified the code to remove checks that don't apply to
  this test.
Removed unused $deviceName local variables.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 158768]