Increase minimum memory required for all distributions. [VDO-5559]
Change 140281 increased the memory requirements for RHEL8 on //eng/vdo
from the general 200MB value to 400MB, presumably based on testing
(though documentation is sparse). This appears to be based on the fact
that a basic pfarm not running VDO uses around 450MB of memory. This
change was not reflected on //eng/linux-vdo, however.

On //eng/linux-vdo, requirements are set at 200MB for all distros. On
RHEL9 this causes the tests to fail occasionally because a vital
process (such as sshd) is killed by the oom-killer. However, inspection
of test logs demonstrates that even when the test passes, it usually
invokes the oom-killer several times. It just gets lucky and only kills
processes that are unimportant or that restart themselves, like
rsyslogd, systemd, or NetworkManager, and which thus do not trigger a
test failure.

It is unclear why the strict memory requirements provoke the oom-killer
on RHEL9 more often that they do on Fedora 37 or 38, which seem to use
a comparable amount of memory. But general memory usage seems to be the
same across all of them, so all expectations should be updated.

Pair:  slegendr

.../VDOTest.pm
Update the minimum memory requirements for all distros from 200MB to
  400MB. This value reflects the memory used by an idle pfarm, and
  allows the tests to complete without invoking the oom-killer.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 158772]