Fix rpm packaging so that multiple rpms are not generated. [VDO-5573]
As it is now, multiple srpms can exist in src/srpms when the tree is 
built due to the srpms being distribution-specific. This causes perl
tests to fail due to them expecting only one match to the 'test -f'
glob pattern. To correct this, the src/srpms directory should be
cleaned each time a build is done, and the srpms themselves should
not be distribution-specific.

Pair: msakai 

.../src/Makefile
Modify the srpms instruction to clean src/srpms prior to building
  new srpms.

.../kvdo.spec
.../vdo.spec
Add new spec files that are not distribution-specific.

.../MANIFEST.yaml
Modify the manifest to use the local spec files instead of the
  ones in src-dist.

[git-p4: depot-paths = "//eng/linux-vdo/": change = 158839]