Replace EXTERNAL_STATIC with STATIC.
It still needs defining as a macro or replacing in the text, but using
a token of the same length as "static" (which is the version we send
upstream) means maintaining indentation (and line breaks) in our
internal code consistent with that which we send upstream will be
easier.

In RPM generation, make sure we only replace "STATIC" as a separate
token at the start of a line.