Hide NULL check for name in uds_resume_index
VDO always knows what device to use and so it always passes it anyway,
so there is no reason for the upstream version to check for a NULL name.

However, since uds_resume_index() is an API method for UDS, it should
still check for a NULL value. Move the NULL handling down into
replace_device(), where it is easy to #ifdef away.

Note that current tests that pass NULL could pass the correct device,
but we should still check that the NULL case is handled properly.