Define test block device interface
Define a new interface to easily get a block_device for test use. Add
getTestBlockDevice() and getTestMultiBlockDevices(), similar to
getTestIndexName() and getTestMultiIndexNames(), and also add
putTestBlockDevice() to free a test device now that they are allocated
objects.

Replace initializerWithIndexName() with initializerWithBlockDevice(),
and update references in the test README file. (The README file is
rather out-of-date and should perhaps be rewritten or removed.)