Reduce build work in overlay script
Build only the perl and stats directories instead of the whole VDO tree.
The perl directory is necessary for the prepare script to work, and the
stats directory is required to properly update generated files.

Omitting the rest of the build has two benefits. First, it saves some
time. Second, it allow the script to process commits which don't build
completely due to missing test or userspace updates.

Signed-off-by: Matthew Sakai <msakai@redhat.com>