dm vdo overlay: various streamlining improvements
Don't include the original commit ID in the commit message because
it will be meaningless (and confusing) upstream.
Replace the version marker with a fixed version corresponding to
our baseline, to prevent spurious diffs.
Automatically sign overlay commits.
Adjust blank lines to attach error handling more closely with the
commands it is checking, and otherwise improve legibility.
Remove handling for COMMIT_FILE variable.

Signed-off-by: Matthew Sakai <msakai@redhat.com>