Disable Fedora 37 testing
With the default builder going from Fedora 37 to Fedora 38, it is not
clear that tests will still be able to build and run on Fedora 37.
However, we also plan to retire Fedora 37 soon, so we can stop running
tests on Fedora 37 now.

Move test builds and commands which only ran on Fedora 37 to run on
Fedora 38 instead. Remove Fedora37 upgrade scenarios and replace them
with Fedora 38 scenarios as appropriate.

Signed-off-by: Matthew Sakai <msakai@redhat.com>