Adapt to the 6.7+ block device interface
Add an alternate block_device opening path that uses bdev_open_by_path()
and bdev_release(), introduced in the 6.7 kernel. Although the older
blkdev_get_by_path() functions appear to still be defined, the
mainline-next kernel does not seem to contain them. The older functions
are slated for eventualy removal, though, so it is better to use the
newer function where they are available.

This change allows vdo-linux to build against the mainline-next kernel.

Signed-off-by: Matthew Sakai <msakai@redhat.com>