Allow fsck exit status 1 (filesystem errors corrected)
Fedora 39 has fsck support for checking the "orphan file" status in
ext4, and will report if the orphan_present flag is set but the orphan
file is empty. It's a reasonable and normal state if file system
updates are interrupted in between metadata blocks, as they are for
rebuild tests, since multiple blocks can't get updated atomically.