[VDOSTORY-283] Replace generated stats files with static files
The stats generation code is somewhat cumbersome, and the generated files
are not easy to keep in sync with upstream, especially when contributors
want to edit them. It is likely simpler from a maintenance point-of-view
to check in static versions for the generated files, and bear the cost of
remembering to keep the various files in sync in the rare cases when we
add or remove a statistics field.

In this change, we added all the files that were generated before in the
repo and updated comments in each file.

We also reflect these changes by updating .gitignore, Makefile and defines.
Stats files are no longer generated, .gitignore will not ignore those
generated files any longer.

Signed-off-by: Chung Chung <cchung@redhat.com>