dm vdo tests: update to use kernel-style mutex functions
Update tests to use the new kernel-style mutex functions now
that all of userspace can use them. However, userspace-only
code is not updated. It would make more sense to directly use
POSIX-style mutexes in userspace code than to add another level
of wrapping.

Also used the kernel-style mutex_destroy(). It might turn out
to be meaningless, but keep it in to be safe for now.

Signed-off-by: Matthew Sakai <msakai@redhat.com>