dm vdo tests: rename internal-only memory-alloc functions
Rename uds_alloc_sprintf to vdo_alloc_sprintf, on the grounds
that we renamed all the other memory allocation functions.
This function is called from vdoFormat, so it will also be
visible in our user code.

Also rename uds_free_const to vdo_free_const to remain
similar to vdo_free. This function is only used in albtest.

Signed-off-by: Matthew Sakai <msakai@redhat.com>