dm vdo tests: fix race condition in BlockMapTreeWrites_t1
CI testing hit this issue, but it's reproducible for me in about
4000 iterations of the test. It is possible for the released vios
to all complete before the write count is checked, causing a
spurious failure. Instead, check for unfinished writes before
releasing the last flusher.

There is at least one more race in this test which will be addressed
later.

Signed-off-by: Matthew Sakai <msakai@redhat.com>