[VDO-5724] Use lvm devices file directly
Changing our code to use the use_devicesfile=0 config option for
lvm commands seems to have broken some other tests that weren't
using config options at all. Rather than adding the
use_devicesfile=0 option everywhere, it makes more sense to
actually use the devices file itself; since this is what LVM
expects us to do. This change makes it so each device in our
stack gets added to the system.devices file in /etc/lvm/devices
right after they are created on the stack and get removed just
before the device is removed from the stack. In this way, all
the devices we care about should be in the devices file.

BlockDevice.pm: Add the device during activation and remove it
during deactivation. This is the same spot the only place in our
code where we used to use the devices file (in ISCSI.pm) did it.

ISCSI.pm: Use the BlockDevice function getLVMDevicesDevicePath()
to make sure we're using the same device name as before for
passthrough devices.

VolumeGroup.pm: Remove the use of use_devicesfile=0 from all
configs. Also add in the config to the removal of thin devices,
since it was missing there.

VDO.pm: Reverse the order of when super::activate gets called so
that the deactivation steps are done in the right order for non
lvm VDO's (like in DoryFastRebuild). In other words, we first
remove the VDO device entry from devices file, then remove
the VDO device itself.

I've left in some logging for now so we can track any issues that
may come up from this change.

Signed-off-by: Bruce Johnston <bjohnsto@redhat.com>