dm: use bio_list_merge_init
This is an adaptation of upstream commit 50bc215030f3 from
Christoph Hellwig. It was part of a series that went in through
linux-block (Jens' tree) so we missed applying it. It's been
in the kernel since 6.10.

Add VDO_USE_ALTERNATE blocks and a userspace definition of
bio_list_merge_init() so that it will build in our lab.

Signed-off-by: Matthew Sakai <msakai@redhat.com>