dm-vdo: Changes are needed to build and test RAWHIDE
RAWHIDE updated some of the system's .h files.
We need to update our .h files to reflect those changes.
Also, removed some of the obsolete inline definitions
that are no longer needed and caused compiler problems.

We also move WARN_ONCE definition from min_heap.h to kernel.h
since it is a better place for it.

Also, we discover that missing-declarations and missing-prototypes
were not turned on by default in gcc before 14.0.1.  Some of our
codes need to be updated since it is now turned on.