Use grubby for kernel option changes on RHEL 9.
It appears that grubby is the new way of updating kernel options, and
has been preferred for a while, but in RHEL 9 there are kernel options
used that are not in the grub config file, making grubby the only
reliable way of updating options without losing others.

Switch only RHEL 9 to grubby, for now.