[VDO-5762] Fix nightly run of DMTest.
We were assuming (incorrectly) that we would only ever be running
dmtest suite using dm-linux source code, which uses a different
set of SRPMS. I've added in the code to deal with running dmtest
using the same SRPM choices that VDOTest.pm does. I've also
updated the make files to allow for dm-linux to still be used.

Signed-off-by: Bruce Johnston <bjohnsto@redhat.com>