dm vdo user: make vdoFillIndex read the VDO geometry
Update vdoFillIndex to read the VDO geometry block to find the
index parameters instead of requiring them as arguments. This
simplifies the interface for vdoFillIndex, and it also removes
the dependency on vdoDumpConfig, which was previously used to
read the geometry parameters.

Also fixed a minor bug where the VDO's bio_offset was not taken
into account. This could only affect VDOs with a converted index.

Signed-off-by: Matthew Sakai <msakai@redhat.com>