[VDO-5793] Add tests to make sure we don't accidentally overwrite VDOs.
VDOManagerTestClobber was a test we previously had when
vdo manager was our main tool for creating and managing
VDO volumes. It consisted of several sub tests that made sure
we didn't overwrite existing VDO volumes on a storage device,
when calling vdoformat or vdo manager commands like
"vdo create".

When we moved to LVM managed VDOs, we removed this test, but
we've decided since then that we would like to bring back
some of the sub tests.

VDOClobber.pm is the test that attempts that to do that. It
uses Unmanaged.pm, which creates VDO volumes directly using
dmsetup create and also calls vdoformat directly as well to
replicate the previous tests features.

Signed-off-by: Bruce Johnston <bjohnsto@redhat.com>