Fix the DMTest/Separated.pm test.
In order to get Separated.pm to work, we will create a single
test machine using DMTest/Grouped.pm. This will allow us to run
the "dmtest list" command in order to get the set of tests
we want to run. Once we have the list, we can create a suite of
tests; one DMTest/Grouped.pm instance for each item in the list.

We also want to update dmtest.suites to make sure the Full suite
uses just vdo tests, and not all tests in dmtest.

Signed-off-by: Bruce Johnston <bjohnsto@redhat.com>