dm vdo: add dmsetup message for returning configuration info
Add a new dmsetup message called config which will return
useful configuration information for the VDO volume and
the uds index associated with it. The output will be in
the form of a YAML string.

The output will also contain a version number to handle
possible future additions to the string.

Signed-off-by: Bruce Johnston <bjohnsto@redhat.com>