dm vdo: change test condition for mainline next repo
We distinguished mainline next and rawhide repo by manually increasing
the PATCHLEVEL by 1 in mainline next kernel src Makefile. This caused
us not to test our code properly.

In rawhide, the generated version.h contains the RHEL_RELEASE_CODE
definition which is not in Fedora and Fedora Next source. We will
use if RHEL_RELEASE_CODE #define condition exists to determine if
a system is mainline next or rawhide.

Signed-off-by: Chung Chung <cchung@redhat.com>