dm vdo: Use VDO_USE_NEXT code with RAWHIDE distro
The linux-next min_heap code is now integrated into RAWHIDE. Add MACRO
to test if a system is RAWHIDE, and use linex-next min_heap code if it is.
We also replaced the VDO_USE_ALTERNATE variable with VDO_USE_NEXT if the
code block is for the linux-next kernel to clarify the usage.

We also reorder code blocks so that the older code comes first and
then the newest code.

Signed-off-by: Chung Chung <cchung@redhat.com>