dm vdo: remove bad check of bi_next field
VDO creates lists of data bios in places but does not for metadata
bios. A sanity check was in place to ensure that no code accidentally
did so for metadata, but erroneously assumed the storage layer below
wouldn't modify the bi_next field while operating on a bio, so a
not-yet-reset bio could have the field set.

Since the check is unnecessary and wrong, remove it.

Signed-off-by: Ken Raeburn <raeburn@redhat.com>