dm vdo: skip corruption check for invalid recovery journal blocks
Checking for corruption should not be done until it has been
determined that the recovery journal block being evaluated is a
valid block, and not one that signals the end of the journal.

Removing the metadata format validation from function
is_exact_recovery_journal_block() is necessary in order to allow
the journal block to be subsequently checked for metadata
corruption, with the appropriate failure code returned when
corruption is detected.

Signed-off-by: Susan LeGendre-McGhee <slegendr@redhat.com>