Add tests for basic recovery and an error path corner case
The corner case involves a journal that has a block with a corrupt
header, as well as a hole due to a bad sequence number. The block
should be identified as invalid, causing us to break out of the
parse_journal_for_recovery() loop. Invalid blocks should not be
checked for metadata corruption.

Signed-off-by: Susan LeGendre-McGhee <slegendr@redhat.com>