[VDO-5749] Add test support for thin pools using VDO.
The second step to supporting a thin pool device using VDO is
to create a test device that supports both VDO features like
getting vdostats output and LVM features like activating and
deactivating the lvm volume. In order to do this, we follow
the same code that our LVMManaged device uses; inheriting
from both LVM.pm and VDO.pm.

Since the new device will act similarly to the LVMManaged
device, we have split LVMManaged.pm into two files;
LVMVDO.pm which contains the double inheritance and other
support functions for all devices that share LVM and VDO,
and a Managed.pm which contains specific code that handles
the stack for LVMManaged.pm.

We also have created a new directory called LVMVDO and
inside of there we have placed two test devices; the
previously mentioned Managed.pm and the new thinpool
using VDO file called ThinPool.pm. Both these inherit
from LVMVDO.pm

In addition, in order to properly handle the new thin
pool device, we need to inherit from LVM/ThinPool.pm.
This creates a triple inheritance but is the cleanest
approach we could come up with.

Finally, because the new device is only supported on
later versions of lvm, we have added a check for the
lvm version in the start of the subtests to skip if
support is not available on the machine the test is
running on.

Signed-off-by: Bruce Johnston <bjohnsto@redhat.com>