dm vdo: Add formatting parameters to the kernel table line
This commit adds the parameters that are currently part of
vdoformat to the vdo table line. There are three parameters;
slab bits, index memory size, and whether the index is sparse
or not. The three parameters will be optional and use defaults
if not passed in. Additional code has been added to do value
checking of these new parameter so we can return errors when
values are incorrect.

The target_type version number has been updated so that clients
of VDO (LVM, Stratis, etc) will know what versions of VDO support
these new parameters and the new formatting code.

Signed-off-by: Bruce Johnston <bjohnsto@redhat.com>