dm vdo user: Add tests for formatting parameters
[VDOSTORY-292] This commit provides tests for the new
table line parameters that will support formatting in
the kernel. It also fixes an issue with the old
VDOFormat.pm test which was succeeding even when it
shouldn't have been.

The first change is to add a flag for deciding what
type of formatting we want to do; the old way calling
vdoformat and the new way formatting directly from the
kernel. Then we create a new function in Unmanaged.pm
to do the formatting in the kernel by wiping the first
block of the storage and then creating a new VDO volume
if parameters were passed into the function. We only
do it for testing. If nothing is passed in for testing,
the normal dmsetup create done in activate will be
used.

Then we add a new test; VDOFormatInKernel.pm to test
the parameters and validation code. This test is
nearly identical to VDOFormat.pm, but has different code
for checking error codes and things since they get
returned from dmsetup and the specific error messages
are written to the kernel log.

Finally, some changes are made to VDOFormat.pm. During
testing, it was noticed that any error message was
accepted because assertExecuteCommand adds the stack
trace to the error message and part of the stack trace
is the call to tryIllegal with the error message match
string. So we changed the search string to be more
explicity by adding vdoformat to the prefix. This makes
the tests only look at the output of vdoformat.

Signed-off-by: Bruce Johnston <bjohnsto@redhat.com>